Updated segment tests
